Costruzione Dao

+++ src/main/java/it/unical/informatica/webapp24/recensioniristoranti/persistenza/dao/postgres/PiattoDaoPostgres.java
@@ -62,7 +62,7 @@ public class PiattoDaoPostgres implements PiattoDao {
                     piatto.setPrezzo(rs.getDouble("p_prezzo"));
 
                 }
-                Long ristId = rs.getLong("r_id");
+                Integer ristId = rs.getInt("r_id");
                 Ristorante r = DBManager.getInstance().getRistoranteDao().findByPrimaryKey(ristId);
                 piatto.addRistorante(r);
